Extract sync environment helpers
